Collisions with ball and block are working
Also, fixed incorrect board width and height.

+++ src/breakout/controller/MoveMessage.java
@@ -5,17 +5,17 @@ package breakout.controller;
  * Honestly, not sure if I should let Ball be moved using MoveMessage as well.
  */
 public class MoveMessage implements Message{
-    private int newCoordinate;
+    private double newCoordinate;
 
     /**
      * Intializes MoveMessage with the new coordinate.
      * @param newCoordinate the new coordinate to update BoardView.
      */
-    public MoveMessage(int newCoordinate) {
+    public MoveMessage(double newCoordinate) {
         this.newCoordinate = newCoordinate;
     }
 
-    public int getNewCoordinate() {
+    public double getNewCoordinate() {
         return newCoordinate;
     }
 }
